Correct return type of MarkerArray::size() + minor spelling correction

+++ kr_viz/rviz_helper/include/rviz_helper/kr_marker.hpp
@@ -63,7 +63,7 @@ class MarkerArray {
 
   Marker &at(int pos);
   const Marker &at(int pos) const;
-  int size() const { return array_.size(); }
+  size_t size() const { return array_.size(); }
 
   //batch marker operation
